feat: merge hooks for auto select
