+++ src/core_modules/capture-core/dataQueries/index.js
@@ -1,1 +1,2 @@
 export { useOrganisationUnit } from './useOrganisationUnit';
+export { useOrgUnitsForAutoSelect } from './useOrgUnitsForAutoSelect';
